Commit Message Test
This is a test to see how the commit messages work, because the git-log's man option was talking about the "title line" of the commit message, and at other places of the "full commit message".
So this is a test to see if we can indeed write full paragraphs in commit messages, and then use this commit to test the different possibilities of the git-log option.
So I guess this will be my last line. Bye.