Finished until ex06
Had a a bit of a hard time with 05 and 06, first for not adding a ', '
at the end of the output, then to make 06 fit into the "25 lines
functions" rule of the 42 norm, but eventually managed to do it without
having to declare sub-fonctions.
I might be realizing right now it does not fit the "80 columns lines"
rule of said norm, but I'm not gonna abort this commit, so it might bz
fixed in the next one.