Finished ex07
Finished the ft_putnbr exercise. Last time during the piscine I had to
put an exception for it to handle the int min. I had to put a 'if it's -2147483648 then print "-2147483648"', which is horrendous, but I couldn't find a way to make my function handle it autonomously. I finally succeeded at that, now if I just change the parameter for a long, it will also handle the long int min, because the algorithm works autonomously without exceptions.